fix: handle polyfills
