+++ frontend/polyfills.js
@@ -1,6 +1,6 @@
 // polyfill of groupBy method
 if (! ('groupBy' in Object)) {
-    Object.prototype.groupBy = function(obj,fn) {
+    Object.prototype.groupBy = function (obj,fn) {
         if (typeof fn !== 'function') throw new Error(`${fn} should be a function`)
         Object.keys(obj).reduce((acc,key) => {
             const group = fn(obj[key])
