+++ frontend/src/main.js
@@ -4,7 +4,7 @@ import { createApp } from 'vue'
 import router from './router'
 import App from './App.vue'
 import { createPinia } from 'pinia'
-import '../polyfills'
+// import '../polyfills'
 
 import {
   Button,
